% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCcontent.R
\name{GCcontent}
\alias{GCcontent}
\title{Calculating GC content}
\usage{
GCcontent(x)
}
\arguments{
\item{x}{A character vector, only consisting of AGC(T/U)}
}
\value{
A character vector
}
\description{
Calculating GC content
}
\examples{
x <- "AGC TTA GCG CTA GC"
GCcontent(x)
}
