% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotide_teller.R
\name{nucleotide_teller}
\alias{nucleotide_teller}
\title{Nucleotide teller}
\usage{
nucleotide_teller(x, nucleotide)
}
\arguments{
\item{x}{A character vector, only consisting of AGC(T/U)}

\item{nucleotide}{What nucleotide to count}
}
\value{
An integer
}
\description{
Nucleotide teller
}
\examples{
x <- "AAA GGG TTT"
nucleotide_teller(x, nucleotide = "A")
}
