% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction_sites.R
\name{restriction_site}
\alias{restriction_site}
\title{Title}
\usage{
restriction_site(x, restriction_sites)
}
\arguments{
\item{x}{A character vector}

\item{restriction_sites}{A character vector}
}
\value{
A data frame
}
\description{
Title
}
\examples{
x <- "AGCTGGGGAAATTAGCTAGGCT"
y <- c("AGCT", "GAAA")
restriction_site(x,y)
}
